fix compressor loop-order
commit dbac106eb6e3cff405e57c0bb6a6dbe3cf307c52 introduced a flaw in
the logic - once a file is downloaded the 'last;' statement exited the
inner loop iterating over the compoentens instead of the outer one
iterating over the compressors.

since both do not inherently depend on each other simply swap them.

noticed while trying to build a pmg-template in a clean environment
with full repository configuration (before the tests worked due to
leftover files in the infodir)

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>