add support to write out apt sources in modern deb822 format
It's a bit cumbersome due to the internal handling being still
targeting the old list format, but good enough for now.

And FWIW, having the apt sources in deb822 support is supported since
a long time (IIRC over 10 years), so we could also just drop support
for the old style and rework the code to something nicer, maybe only
after bumping and uploading this package for the next major Debian
release (trixie).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>